# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food API",
        "description": "API for managing dishes, categories, ingredients, and AI-powered dietary recommendations.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/recommendations/analyze/{plate_id}": {
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Analyze a dish against the user's dietary restrictions using AI",
                "operationId": "7ebd2857f6f0f39b4b45eae0b3447f96",
                "parameters": [
                    {
                        "name": "plate_id",
                        "in": "path",
                        "description": "ID of the dish to analyze",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "dish_id": {
                                            "type": "integer"
                                        },
                                        "score": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "label": {
                                            "type": "string",
                                            "example": "Highly Recommended"
                                        },
                                        "warning_message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ready"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dish not found"
                    },
                    "500": {
                        "description": "Invalid AI response"
                    },
                    "502": {
                        "description": "Gemini API error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "List all recommendations for the authenticated user",
                "operationId": "b50096a7b015b6eb5e37f889d4d86223",
                "responses": {
                    "200": {
                        "description": "List of recommendations with dish info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "score": {
                                                "type": "integer"
                                            },
                                            "label": {
                                                "type": "string"
                                            },
                                            "warning_message": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "dish": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recommendations/{plate_id}": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get a single recommendation for a dish",
                "operationId": "384cb70199c20e4b29cd490f1eb446c4",
                "parameters": [
                    {
                        "name": "plate_id",
                        "in": "path",
                        "description": "ID of the dish",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation details"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user info",
                "operationId": "1d8a3a3c4101a12de97830e80b50d7a3",
                "responses": {
                    "200": {
                        "description": "User info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user dietary profile",
                "operationId": "9a6644419426815611b40315622b6c75",
                "responses": {
                    "200": {
                        "description": "Dietary profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "dietary_tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "vegan",
                                                "gluten_free"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update authenticated user dietary profile",
                "operationId": "2c8c78a7ec16d5df7eab9b1005fcd724",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dietary_tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "vegan",
                                            "no_sugar"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated dietary tags",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "dietary_tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}